feat(proxy): add support for proxy chains and enhance QUIC connection handling
